Move message definition to separate package to prevent import cycle
