+++ relayer/message/message_processors_test.go
@@ -1,4 +1,4 @@
-package relayer
+package message
 
 import (
 	"math/big"
