+++ relayer/message/message_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package relayer
+package message
 
 import (
 	"math/big"
